+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## unreleased
 
 - expose `Path_condition.pp`
+- substitute symbol=value equalities in the PC
 
 ## 0.1 - 2026-02-09